refactor(test): reorganize download_history test files
- Moved JSON serialization tests to new json_serializer_test.go
- Updated build tags in concurrent_test.go to use 'race' tag
- Modified Makefile to handle test tags more explicitly
- Cleaned up imports and test organization in download_history_test.go
- Improved test isolation and maintainability
