+++ Makefile
@@ -2,7 +2,7 @@
 RM := /bin/rm -rf
 MKDIR := /bin/mkdir -p
 GO_BUILD := go build
-GO_TEST := go test -tags test
+GO_TEST := go test
 
 # Declare phony targets that don't produce files
 .PHONY: build test test-unit test-full test-race clean run pre-commit
@@ -18,15 +18,15 @@ test: test-unit
 
 # Run unit tests (excludes integration tests)
 test-unit:
-	$(GO_TEST) ./...
+	$(GO_TEST) -tags test ./...
 
 # Run integration tests (requires Synology NAS environment variables)
 test-integration:
-	$(GO_TEST) -tags=integration -count=1 ./synology_drive_api/...
+	$(GO_TEST) -tags test,integration -count=1 ./synology_drive_api/...
 
 # Run race tests (to find race conditions)
 test-race:
-	$(GO_TEST) -tags=test -race -v -timeout=5s ./download_history/...
+	$(GO_TEST) -tags test,race -race -v -timeout=5s -count=1 ./download_history/...
 
 # Run all tests (unit + integration + race)
 test-full: test-unit test-integration test-race
